requirements.txt
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
certifi==2025.11.12
charset-normalizer==3.4.4
click @ file:///D:/bld/bld/rattler-build_click_1764518654/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1755674618058/work
google_search_results==2.4.2
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
MarkupSafe @ file:///D:/bld/markupsafe_1759055175616/work
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
pandas @ file:///D:/bld/pandas_1764614482805/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1761503229/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
RapidFuzz @ file:///D:/bld/rapidfuzz_1762522562398/work
requests==2.32.5
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1765719872007/work
urllib3==2.6.2
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_werkzeug_1764423417/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
